requirements.txt
quart
hypercorn
python-dotenv
supabase
asyncpg
numpy
openai
langchain
langchain-openai
langchain-groq
langchain-community
langchain-core
langchain-text-splitters
langgraph
google-generativeai
aiohttp
pydantic

# Development dependencies
pre-commit
